Rename to fanoutprocessor
