+++ exporter/zipkinexporter/zipkin_test.go
@@ -33,7 +33,7 @@ import (
 
 	"github.com/open-telemetry/opentelemetry-service/internal/config/viperutils"
 	"github.com/open-telemetry/opentelemetry-service/internal/testutils"
-	"github.com/open-telemetry/opentelemetry-service/processor/multiprocessor"
+	"github.com/open-telemetry/opentelemetry-service/processor/fanoutprocessor"
 	"github.com/open-telemetry/opentelemetry-service/receiver/receivertest"
 	"github.com/open-telemetry/opentelemetry-service/receiver/zipkinreceiver"
 )
@@ -157,7 +157,7 @@ zipkin:
 	tes[0].(*zipkinExporter).reporter = mzr
 
 	// Run the Zipkin receiver to "receive spans upload from a client application"
-	zexp := multiprocessor.NewTraceProcessor(tes)
+	zexp := fanoutprocessor.NewTraceProcessor(tes)
 	zi, err := zipkinreceiver.New(":0", zexp)
 	if err != nil {
 		t.Fatalf("Failed to create a new Zipkin receiver: %v", err)
